beautify json results
