+++ src/index.css
@@ -47,6 +47,10 @@
     @apply flex flex-col items-center p-5 mx-5 rounded-lg overflow-x-auto;
   }
 
+  .result-message {
+    @apply mb-8 px-4 py-2 border border-borders bg-tertiary;
+  }
+
   .form-area {
     @apply flex flex-col items-center max-w-sm p-4 my-0 mb-2 mx-auto rounded-lg bg-blank;
   }
